+++ src/main/java/blackjack/card/Card.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card;
+package blackjack.card;
 
 public abstract class Card {
