+++ src/main/java/blackjack/card/CardRank.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card;
+package blackjack.card;
 
 public enum CardRank {
