+++ src/main/java/blackjack/card/CardSide.java
@@ -1,4 +1,4 @@
-package blackjack.domain.card;
+package blackjack.card;
 
 public enum CardSide {
