+++ src/test/java/blackjack/participant/NicknameTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain.gamer;
+package blackjack.participant;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
